restore single file
